+++ .semaphore/publish-or-perish.sh
@@ -6,12 +6,18 @@
 
 set -efuo pipefail
 
-git fetch --tags
-TAG=v$(python setup.py --version)
+sudo pip install toml
+TAG=v$(grep -Po '__version__ = .\K[0-9\\.]+' opnieuw/__init__.py)
 # If the tag already exist this command will fail and the job will exit
 # without raising an error. Otherwise, we will build the project,
 # publish to PyPi, and push the tag to github.
+git fetch --tags
 git tag "$TAG" || exit 0
+
+sudo pip install build
+python3 -m build
+
 sudo pip install twine
-twine upload --skip-existing dist/*
+python3 -m twine upload --skip-existing dist/*
+
 git push origin "$TAG"