+++ opnieuw/retries.py
@@ -14,13 +14,13 @@ import logging
 import random
 import sys
 import time
-from types import CoroutineType
+import typing_extensions
 import warnings
 from collections import defaultdict
 from collections.abc import Callable, Iterator
 from contextlib import contextmanager
 from contextvars import ContextVar
-from typing import TypeVar, Awaitable, Any, Coroutine, cast, overload
+from typing import TypeVar, Awaitable, cast, overload
 
 if sys.version_info < (3, 10):
     from typing_extensions import ParamSpec
@@ -297,4 +297,24 @@ def retry(
 # We expose `retry_async` for backwards-compatibility.
 # However, nowadays the main `retry` decorator is able
 # to accept both sync and async functions directly.
-retry_async = retry
+@typing_extensions.deprecated("Use the normal `retry` instead, as works for both sync and async functions", category=None)
+def retry_async(
+    *,
+    retry_on_exceptions: type[Exception] | tuple[type[Exception], ...],
+    max_calls_total: int = 3,
+    retry_window_after_first_call_in_seconds: int = 60,
+    namespace: str | None = None,
+) -> Callable[[Callable[P, Awaitable[R]]], Callable[P, Awaitable[R]]]:
+    """
+    Functionally the same as `retry`.
+
+    In the past, async functions required using `retry_async`,
+    but nowadays the main `retry` decorator can be used
+    for both sync and async functions.
+    """
+    return retry(
+        retry_on_exceptions=retry_on_exceptions, 
+        max_calls_total=max_calls_total,
+        retry_window_after_first_call_in_seconds=retry_window_after_first_call_in_seconds,
+        namespace=namespace
+    )