+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----
 **Release highlights:**
 
+
 3.2.0
 -----